requirements.txt
influxdb==5.3.0
requests==2.24.0
python-aqi==0.6.1